admin/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
PyJWT==2.3.0
python-dotenv==0.19.0
pymongo==3.12.0
pyyaml==6.0
Flask-Cors==4.0.0 